requirements.txt
Django
dj-database-url
djangorestframework
django-cors-headers
django-filter
psycopg2-binary
whitenoise
numpy
attrs
